Fix a bug related to the 'dirname' attribute.
- When using the dirname attribute in the 'dirtt' tag
the tree was not properly created, if the 'dirname' attribute
was used in another 'dir' tag. In order to fix the bug, and
avoid further bugs related to this issue, tree traversal is
stacked up. This means whenever we find a 'dirtt' or a 'dir'
tag, the current directory is pushed down to a stack and
when a the corresponding closing tag is found we 'chdir'
to the element on top of the stack.

- A test case was added to make sure the dirname attribute
is properly used.